faas-cli/pkg/requirements.txt
joblib==1.1.0
numpy==1.21.6
Pillow==9.2.0
PyMySQL==1.0.2
scikit_learn==1.1.2
